\name{findFasit}
\alias{findFasit}
\title{Search through a FASIT df}
\usage{
findFasit(fdf, params)
}
\arguments{
  \item{fdf}{A FASIT data frame}

  \item{params}{A key-value list of column names and search
  criteria (currently only additive search is implemented)}
}
\description{
Search through a FASIT df
}

